refactor: revert Builder -> Compiler internal name change
