+++ internal/backend/bw6-761/cs/r1cs_test.go
@@ -40,7 +40,7 @@ func TestSerialization(t *testing.T) {
 				return
 			}
 
-			r1cs1, err := frontend.Compile(ecc.BW6_761, r1cs.NewCompiler, tc.Circuit)
+			r1cs1, err := frontend.Compile(ecc.BW6_761, r1cs.NewBuilder, tc.Circuit)
 			if err != nil {
 				t.Fatal(err)
 			}
@@ -49,7 +49,7 @@ func TestSerialization(t *testing.T) {
 			}
 
 			// copmpile a second time to ensure determinism
-			r1cs2, err := frontend.Compile(ecc.BW6_761, r1cs.NewCompiler, tc.Circuit)
+			r1cs2, err := frontend.Compile(ecc.BW6_761, r1cs.NewBuilder, tc.Circuit)
 			if err != nil {
 				t.Fatal(err)
 			}
@@ -138,7 +138,7 @@ func (circuit *circuit) Define(api frontend.API) error {
 func BenchmarkSolve(b *testing.B) {
 
 	var c circuit
-	ccs, err := frontend.Compile(ecc.BW6_761, r1cs.NewCompiler, &c)
+	ccs, err := frontend.Compile(ecc.BW6_761, r1cs.NewBuilder, &c)
 	if err != nil {
 		b.Fatal(err)
 	}
