+++ internal/generator/backend/template/representations/tests/r1cs.go.tmpl
@@ -24,7 +24,7 @@ func TestSerialization(t *testing.T) {
 			}
 		{{end}}
 
-		r1cs1, err := frontend.Compile(ecc.{{ .CurveID }}, r1cs.NewCompiler,	tc.Circuit)
+		r1cs1, err := frontend.Compile(ecc.{{ .CurveID }}, r1cs.NewBuilder,	tc.Circuit)
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -33,7 +33,7 @@ func TestSerialization(t *testing.T) {
 		}
 
 		// copmpile a second time to ensure determinism
-		r1cs2, err := frontend.Compile(ecc.{{ .CurveID }},r1cs.NewCompiler, tc.Circuit)
+		r1cs2, err := frontend.Compile(ecc.{{ .CurveID }},r1cs.NewBuilder, tc.Circuit)
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -123,7 +123,7 @@ func (circuit *circuit) Define(api frontend.API) error {
 func BenchmarkSolve(b *testing.B) {
 
 	var c circuit 
-	ccs, err := frontend.Compile(ecc.{{ .CurveID }},r1cs.NewCompiler, &c)
+	ccs, err := frontend.Compile(ecc.{{ .CurveID }},r1cs.NewBuilder, &c)
 	if err != nil {
 		b.Fatal(err)
 	}
