+++ std/algebra/sw_bls12377/g2_test.go
@@ -272,7 +272,7 @@ func BenchmarkDoubleAffineG2(b *testing.B) {
 	var c g2DoubleAffine
 	b.Run("groth16", func(b *testing.B) {
 		for i := 0; i < b.N; i++ {
-			ccsBench, _ = frontend.Compile(ecc.BW6_761, r1cs.NewCompiler, &c)
+			ccsBench, _ = frontend.Compile(ecc.BW6_761, r1cs.NewBuilder, &c)
 		}
 
 	})
@@ -283,7 +283,7 @@ func BenchmarkAddAssignAffineG2(b *testing.B) {
 	var c g2AddAssignAffine
 	b.Run("groth16", func(b *testing.B) {
 		for i := 0; i < b.N; i++ {
-			ccsBench, _ = frontend.Compile(ecc.BW6_761, r1cs.NewCompiler, &c)
+			ccsBench, _ = frontend.Compile(ecc.BW6_761, r1cs.NewBuilder, &c)
 		}
 
 	})
@@ -294,7 +294,7 @@ func BenchmarkDoubleAndAddAffineG2(b *testing.B) {
 	var c g2DoubleAndAddAffine
 	b.Run("groth16", func(b *testing.B) {
 		for i := 0; i < b.N; i++ {
-			ccsBench, _ = frontend.Compile(ecc.BW6_761, r1cs.NewCompiler, &c)
+			ccsBench, _ = frontend.Compile(ecc.BW6_761, r1cs.NewBuilder, &c)
 		}
 
 	})
