+++ std/algebra/sw_bls24315/pairing_test.go
@@ -214,6 +214,6 @@ func TestTriplePairingBLS24315(t *testing.T) {
 
 func BenchmarkPairing(b *testing.B) {
 	var c pairingBLS24315
-	ccsBench, _ = frontend.Compile(ecc.BW6_633, r1cs.NewCompiler, &c)
+	ccsBench, _ = frontend.Compile(ecc.BW6_633, r1cs.NewBuilder, &c)
 	b.Log("groth16", ccsBench.GetNbConstraints())
 }
